docs(examples): add missing step for building at repository root first
This adds a step to some of the example README files for installing
dependencies and building at the repository root first.

I was running into the same issue as #513 because I was following the
instructions exactly, not realizing that I needed to first install
dependencies and build at the repository root.

Fixes #513.